Added "Knight Speed" option
You can now change the movement speed of Knights: 0.7x, 0.85x, 1x, 1.15x, 1.3x. It may just be me, but movement looks kinda jittery when high-speed knights change direction. I guess it's good enough for now.

+++ Trogdor-Reburninated/src/general.h
@@ -21,7 +21,7 @@ class FrameState {
 		void increment();
 		bool startingFrame(Uint16);
 		bool endingFrame(Uint16);
-		void subtract(Uint8);
+		void subtract(double);
 		void mod(Uint8);
 };
